# fix: reduced ReportAvatarPosition overhead