--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarReporterController/Implementation/AvatarReporterController.cs
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarReporterController/Implementation/AvatarReporterController.cs
@@ -6,6 +6,9 @@ using DCL.Controllers;
 using DCL.Helpers;
 using UnityEngine;
 
+/// <summary>
+/// This controller checks if an avatar entered or exited a scene by checking its position every frame
+/// </summary>
 public class AvatarReporterController : IAvatarReporterController
 {
     private string avatarId;
@@ -36,15 +39,8 @@ public class AvatarReporterController : IAvatarReporterController
     
     string GetcurrentSceneIdNonAlloc(Vector2Int coords)
     {
-        foreach (KeyValuePair<string, IParcelScene> parcelScene in worldState.loadedScenes)
-        {
-            var parcels = parcelScene.Value.sceneData.parcels;
-
-            if (parcels != null && parcels.Contains(coords))
-            {
-                return parcelScene.Key;
-            }
-        }
+        if (worldState.loadedScenesByCoordinate.ContainsKey(coords))
+            return worldState.loadedScenesByCoordinate[coords];
         
         return null;
     }