--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarReporterController/Tests/AvatarReporterControllerShould.cs
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarReporterController/Tests/AvatarReporterControllerShould.cs
@@ -12,6 +12,7 @@ public class AvatarReporterControllerShould
     private IAvatarReporterController reporterController;
     private IWorldState worldState;
     private Dictionary<string, IParcelScene> scenes;
+    private Dictionary<Vector2Int, string> sceneByCoords;
 
     [SetUp]
     public void SetUp()
@@ -29,8 +30,10 @@ public class AvatarReporterControllerShould
         });
 
         scenes = new Dictionary<string, IParcelScene>() { { "scene0", scene0 } };
+        sceneByCoords = new Dictionary<Vector2Int, string> { { Vector2Int.zero, "scene0" } };
 
         worldState.loadedScenes.Returns(scenes);
+        worldState.loadedScenesByCoordinate.Returns(sceneByCoords);
     }
 
     [Test]
@@ -96,16 +99,21 @@ public class AvatarReporterControllerShould
         reporterController.reporter.Received(1).ReportAvatarSceneChange("1", null);
 
         var loadScene = Substitute.For<IParcelScene>();
+        Vector2Int scenePos = new Vector2Int(1, 0);
+
+        string sceneId = "sceneLoaded";
+
         loadScene.sceneData.Returns(new LoadParcelScenesMessage.UnityParcelScene()
         {
-            id = "sceneLoaded",
-            parcels = new[] { new Vector2Int(1, 0) }
+            id = sceneId,
+            parcels = new[] { scenePos }
         });
 
-        scenes.Add("sceneLoaded", loadScene);
+        scenes.Add(sceneId, loadScene);
+        sceneByCoords.Add(scenePos, sceneId);
 
         reporterController.ReportAvatarPosition(position);
-        reporterController.reporter.Received(1).ReportAvatarSceneChange("1", "sceneLoaded");
+        reporterController.reporter.Received(1).ReportAvatarSceneChange("1", sceneId);
     }
 
     [Test]
@@ -116,13 +124,16 @@ public class AvatarReporterControllerShould
         reporterController.reporter.Received(1).ReportAvatarSceneChange("1", "scene0");
 
         var loadScene = Substitute.For<IParcelScene>();
+        Vector2Int position = new Vector2Int(1, 0);
+
         loadScene.sceneData.Returns(new LoadParcelScenesMessage.UnityParcelScene()
         {
             id = "sceneLoaded",
-            parcels = new[] { new Vector2Int(1, 0) }
+            parcels = new[] { position }
         });
 
         scenes.Add("sceneLoaded", loadScene);
+        sceneByCoords.Add(position, "sceneLoaded");
 
         reporterController.ReportAvatarPosition(new Vector3(ParcelSettings.PARCEL_SIZE, 0, 0));
         reporterController.reporter.Received(1).ReportAvatarSceneChange("1", "sceneLoaded");