--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarShape.cs
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarShape.cs
@@ -110,6 +110,8 @@ namespace DCL
         {
             playerName = GetComponentInChildren<IPlayerName>();
             playerName?.Hide(true);
+            
+            Environment.i.platform.updateEventHandler.AddListener(IUpdateEventHandler.EventType.Update, OnUpdate);
         }
 
         private void PlayerClicked()
@@ -121,6 +123,8 @@ namespace DCL
 
         public void OnDestroy()
         {
+            Environment.i.platform.updateEventHandler.RemoveListener(IUpdateEventHandler.EventType.Update, OnUpdate);
+
             Cleanup();
 
             if (poolableObject != null && poolableObject.isInsidePool)
@@ -301,7 +305,7 @@ namespace DCL
                 player.playerName.SetName(model.name);
         }
 
-        private void Update()
+        private void OnUpdate()
         {
             if (player != null)
             {