--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/WorldState.cs
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/WorldState.cs
@@ -11,6 +11,7 @@ namespace DCL
     {
         public HashSet<string> readyScenes { get; set; } = new HashSet<string>();
         public Dictionary<string, IParcelScene> loadedScenes { get; set; } = new Dictionary<string, IParcelScene>();
+        public Dictionary<Vector2Int, string> loadedScenesByCoordinate { get; set; } = new Dictionary<Vector2Int, string>();
         public List<IParcelScene> scenesSortedByDistance { get; set; } = new List<IParcelScene>();
         public List<string> globalSceneIds { get; set; } = new List<string>();
         public string currentSceneId { get; set; } = null;