## What does this PR change?

`AvatarReporterController.ReportAvatarPosition` had a small overhead and allocations, those are gone.

Fixes #2701 

## How to test the changes?

Scenes with `onEnterSceneObservable` and `onLeaveSceneObservable` should work properly. 

## Our Code Review Standards

https://github.com/decentraland/unity-renderer/blob/master/docs/code-review-standards.md

## Performance Data

# OLD
67 players (-100,126 ) 30 seconds WITHOUT UI (pressing U)
```
 * PERFORMANCE SCORE (0-100) -> 10
 * lowest frame time -> 61.7ms
 * average frame time -> 77.4ms
 * highest frame time -> 100.0ms
 * 1 percentile frame time -> 64.3ms
 * 50 percentile frame time -> 74.4ms
 * 99 percentile frame time -> 100.0ms
 * error percentage -> ±1.3%
 * total hiccups (>0.05ms frames) -> 387 (99.74% of frames were hiccups)
 * total hiccups time (seconds) -> 29.94901
 * total frames -> 388
 * total frames time (seconds) -> 30.0292
```

# NEW
67 players (-100,126 ) 30 seconds WITHOUT UI (pressing U)
```
 * PERFORMANCE SCORE (0-100) -> 11
 * lowest frame time -> 60.8ms
 * average frame time -> 74.5ms
 * highest frame time -> 100.0ms
 * 1 percentile frame time -> 62.2ms
 * 50 percentile frame time -> 70.6ms
 * 99 percentile frame time -> 100.0ms
 * error percentage -> ±1.5%
 * total hiccups (>0.05ms frames) -> 402 (99.75% of frames were hiccups)
 * total hiccups time (seconds) -> 29.94498
 * total frames -> 403
 * total frames time (seconds) -> 30.0436
```